+++ app/src/main/java/org/akvo/caddisfly/ui/PreferencesDeveloperFragment.java
@@ -1,0 +1,49 @@
+package org.akvo.caddisfly.ui;
+
+
+import android.app.Fragment;
+import android.app.FragmentTransaction;
+import android.os.Bundle;
+import android.preference.Preference;
+import android.preference.PreferenceFragment;
+import android.support.annotation.NonNull;
+import android.view.LayoutInflater;
+import android.view.View;
+import android.view.ViewGroup;
+
+import org.akvo.caddisfly.R;
+
+/**
+ * A simple {@link Fragment} subclass.
+ */
+public class PreferencesDeveloperFragment extends PreferenceFragment {
+
+    public PreferencesDeveloperFragment() {
+        // Required empty public constructor
+    }
+
+    @Override
+    public void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        addPreferencesFromResource(R.xml.pref_developer);
+    }
+
+    @Override
+    public View onCreateView(@NonNull LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState) {
+        // Inflate the layout for this fragment
+        View rootView = inflater.inflate(R.layout.card_row, container, false);
+
+        Preference cameraPreviewPreference = findPreference("cameraPreview");
+        if (cameraPreviewPreference != null) {
+            cameraPreviewPreference.setOnPreferenceClickListener(new Preference.OnPreferenceClickListener() {
+                public boolean onPreferenceClick(Preference preference) {
+                    final FragmentTransaction ft = getFragmentManager().beginTransaction();
+                    CameraFragment cameraFragment = CameraFragment.newInstance(true);
+                    cameraFragment.show(ft, "cameraFragment");
+                    return true;
+                }
+            });
+        }
+        return rootView;
+    }
+}