+++ app/src/main/java/org/akvo/caddisfly/ui/PreferencesGeneralFragment.java
@@ -1,0 +1,34 @@
+package org.akvo.caddisfly.ui;
+
+
+import android.app.Fragment;
+import android.os.Bundle;
+import android.preference.PreferenceFragment;
+import android.support.annotation.NonNull;
+import android.view.LayoutInflater;
+import android.view.View;
+import android.view.ViewGroup;
+
+import org.akvo.caddisfly.R;
+
+/**
+ * A simple {@link Fragment} subclass.
+ */
+public class PreferencesGeneralFragment extends PreferenceFragment {
+
+    public PreferencesGeneralFragment() {
+        // Required empty public constructor
+    }
+
+    @Override
+    public void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        addPreferencesFromResource(R.xml.pref_general);
+    }
+
+    @Override
+    public View onCreateView(@NonNull LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState) {
+        return inflater.inflate(R.layout.card_row, container, false);
+    }
+
+}