+++ app/src/main/java/org/akvo/caddisfly/ui/SettingsActivity.java
@@ -16,17 +16,11 @@
 
 package org.akvo.caddisfly.ui;
 
-import android.annotation.TargetApi;
-import android.app.FragmentTransaction;
+import android.app.Activity;
 import android.content.Context;
 import android.content.Intent;
 import android.content.SharedPreferences;
-import android.content.res.Configuration;
-import android.os.Build;
 import android.os.Bundle;
-import android.preference.Preference;
-import android.preference.PreferenceActivity;
-import android.preference.PreferenceFragment;
 import android.preference.PreferenceManager;
 import android.support.annotation.Nullable;
 import android.support.v7.widget.Toolbar;
@@ -36,99 +30,19 @@ import android.view.View;
 import android.widget.LinearLayout;
 
 import org.akvo.caddisfly.R;
-import org.akvo.caddisfly.app.MainApp;
 import org.akvo.caddisfly.util.PreferencesUtils;
-import org.akvo.caddisfly.util.UpdateCheckTask;
 
-import java.util.List;
-
-/**
- * A {@link android.preference.PreferenceActivity} that presents a set of application settings. On
- * handset devices, settings are presented as a single list. On tablets,
- * settings are split by category, with category headers shown to the left of
- * the list of settings.
- * <p/>
- * See <a href="http://developer.android.com/design/patterns/settings.html">
- * Android Design: Settings</a> for design guidelines and the <a
- * href="http://developer.android.com/guide/topics/ui/settings.html">Settings
- * API Guide</a> for more information on developing a Settings UI.
- */
-public class SettingsActivity extends PreferenceActivity
+public class SettingsActivity extends Activity
         implements SharedPreferences.OnSharedPreferenceChangeListener {
-    /**
-     * Determines whether to always show the simplified settings UI, where
-     * settings are presented in a single list. When false, settings are shown
-     * as a master/detail two-pane view on tablets. When true, a single pane is
-     * shown on tablets.
-     */
-    private static final boolean ALWAYS_SIMPLE_PREFS = false;
-    /**
-     * A preference value change listener that updates the preference's summary
-     * to reflect its new value.
-     */
-    private static Preference.OnPreferenceChangeListener sBindPreferenceSummaryToValueListener = new Preference.OnPreferenceChangeListener() {
-        @Override
-        public boolean onPreferenceChange(Preference preference, Object value) {
-            String stringValue = value.toString();
-
-            // For all other preferences, set the summary to the value's
-            // simple string representation.
-            preference.setSummary(stringValue);
-            return true;
-        }
-    };
-
-    /**
-     * Helper method to determine if the device has an extra-large screen. For
-     * example, 10" tablets are extra-large.
-     */
-    private static boolean isXLargeTablet(Context context) {
-        return (context.getResources().getConfiguration().screenLayout
-                & Configuration.SCREENLAYOUT_SIZE_MASK) >= Configuration.SCREENLAYOUT_SIZE_XLARGE;
-    }
-
-    /**
-     * Determines whether the simplified settings UI should be shown. This is
-     * true if this is forced via {@link #ALWAYS_SIMPLE_PREFS}, or the device
-     * doesn't have newer APIs like {@link android.preference.PreferenceFragment}, or the device
-     * doesn't have an extra-large screen. In these cases, a single-pane
-     * "simplified" settings UI should be shown.
-     */
-    private static boolean isSimplePreferences(Context context) {
-        return Build.VERSION.SDK_INT < Build.VERSION_CODES.HONEYCOMB
-                || !isXLargeTablet(context);
-    }
-
-    /**
-     * Binds a preference's summary to its value. More specifically, when the
-     * preference's value is changed, its summary (line of text below the
-     * preference title) is updated to reflect the value. The summary is also
-     * immediately updated upon calling this method. The exact display format is
-     * dependent on the type of preference.
-     *
-     * @see #sBindPreferenceSummaryToValueListener
-     */
-    private static void bindPreferenceSummaryToValue(Preference preference) {
-        // Set the listener to watch for value changes.
-        preference.setOnPreferenceChangeListener(sBindPreferenceSummaryToValueListener);
-
-        // Trigger the listener immediately with the preference's
-        // current value.
-        sBindPreferenceSummaryToValueListener.onPreferenceChange(preference,
-                PreferenceManager
-                        .getDefaultSharedPreferences(preference.getContext())
-                        .getString(preference.getKey(), ""));
-    }
 
     @Override
     protected void onPostCreate(Bundle savedInstanceState) {
         super.onPostCreate(savedInstanceState);
-        setupSimplePreferencesScreen();
 
-        // get the root container of the preferences list
-        LinearLayout root = (LinearLayout) findViewById(android.R.id.list).getParent().getParent().getParent();
+        //Add a tool bar on this settings screen
+        LinearLayout root = (LinearLayout) findViewById(android.R.id.list).getParent().getParent().getParent().getParent().getParent();
         Toolbar bar = (Toolbar) LayoutInflater.from(this).inflate(R.layout.preferences_toolbar, root, false);
-        root.addView(bar, 0); // insert at top
+        root.addView(bar, 0);
         bar.setNavigationOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
@@ -137,134 +51,48 @@ public class SettingsActivity extends PreferenceActivity
         });
     }
 
-    /**
-     * Shows the simplified settings UI if the device configuration if the
-     * device configuration dictates that a simplified, single-pane UI should be
-     * shown.
-     */
-    private void setupSimplePreferencesScreen() {
-        if (!isSimplePreferences(this)) {
-            return;
-        }
-
-        // In the simplified UI, fragments are not used at all and we instead
-        // use the older PreferenceActivity APIs.
-
-        // Add 'general' preferences.
-        addPreferencesFromResource(R.xml.pref_general);
-
-        boolean developerMode = PreferencesUtils.getBoolean(this, R.string.developerModeKey, false);
-        if (developerMode) {
-            addPreferencesFromResource(R.xml.pref_developer);
-        }
-
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    @Override
-    public boolean onIsMultiPane() {
-        return isXLargeTablet(this) && !isSimplePreferences(this);
-    }
-
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
+
+        setContentView(R.layout.activity_settings);
+
         PreferenceManager.getDefaultSharedPreferences(this.getApplicationContext())
                 .registerOnSharedPreferenceChangeListener(this);
 
+        getFragmentManager().beginTransaction()
+                .replace(R.id.content, new PreferencesGeneralFragment())
+                .commit();
+
+        getFragmentManager().beginTransaction()
+                .replace(R.id.content2, new PreferencesOtherFragment())
+                .commit();
+
+        boolean developerMode = PreferencesUtils.getBoolean(this, R.string.developerModeKey, false);
+        if (developerMode) {
+            getFragmentManager().beginTransaction()
+                    .add(R.id.content3, new PreferencesDeveloperFragment())
+                    .commit();
+        }
     }
 
     @Nullable
     @Override
     public View onCreateView(String name, Context context, AttributeSet attrs) {
-        Preference aboutPreference = findPreference("about");
-        if (aboutPreference != null) {
-            aboutPreference.setOnPreferenceClickListener(new Preference.OnPreferenceClickListener() {
-                public boolean onPreferenceClick(Preference preference) {
-                    final Intent intent = new Intent(getBaseContext(), AboutActivity.class);
-                    startActivity(intent);
-                    return true;
-                }
-            });
+
+        if (getIntent().getBooleanExtra("calibrate", false)) {
+            final Intent intent = new Intent(getBaseContext(), CalibrateListActivity.class);
+            startActivity(intent);
         }
-
-        Preference updatePreference = findPreference("checkUpdate");
-        if (updatePreference != null) {
-            updatePreference.setOnPreferenceClickListener(new Preference.OnPreferenceClickListener() {
-                public boolean onPreferenceClick(Preference preference) {
-                    checkUpdate(false);
-                    return true;
-                }
-            });
-        }
-
-        Preference calibratePreference = findPreference("calibrate");
-        if (calibratePreference != null) {
-            calibratePreference.setOnPreferenceClickListener(new Preference.OnPreferenceClickListener() {
-                public boolean onPreferenceClick(Preference preference) {
-                    final Intent intent = new Intent(getBaseContext(), CalibrateListActivity.class);
-                    startActivity(intent);
-                    return true;
-                }
-            });
-        }
-
-        Preference cameraPreviewPreference = findPreference("cameraPreview");
-        if (cameraPreviewPreference != null) {
-            cameraPreviewPreference.setOnPreferenceClickListener(new Preference.OnPreferenceClickListener() {
-                public boolean onPreferenceClick(Preference preference) {
-                    final FragmentTransaction ft = getFragmentManager().beginTransaction();
-                    CameraFragment cameraFragment = CameraFragment.newInstance(true);
-                    cameraFragment.show(ft, "cameraFragment");
-                    return true;
-                }
-            });
-        }
-
-
         return super.onCreateView(name, context, attrs);
-    }
-
-    void checkUpdate(boolean background) {
-        UpdateCheckTask updateCheckTask = new UpdateCheckTask(this, background, MainApp.getVersion(this));
-        updateCheckTask.execute();
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    @Override
-    @TargetApi(Build.VERSION_CODES.HONEYCOMB)
-    public void onBuildHeaders(List<Header> target) {
-        if (!isSimplePreferences(this)) {
-            loadHeadersFromResource(R.xml.pref_headers, target);
-        }
     }
 
     @Override
     public void onSharedPreferenceChanged(SharedPreferences sharedPreferences, String s) {
         Context context = this.getApplicationContext();
         assert context != null;
-        //noinspection CallToStringEquals
+
         if (context.getString(R.string.languageKey).equals(s)) {
-
-            PreferenceManager.getDefaultSharedPreferences(this.getApplicationContext())
-                    .unregisterOnSharedPreferenceChangeListener(this);
-
-//                Locale myLocale = new Locale(
-//                        sharedPreferences.getString(s, Config.DEFAULT_LOCALE));
-//                Resources res = this.getResources();
-//                DisplayMetrics dm = res.getDisplayMetrics();
-//                Configuration conf = res.getConfiguration();
-//                conf.locale = myLocale;
-//                if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.JELLY_BEAN_MR1) {
-//                    conf.setLayoutDirection(myLocale);
-//                }
-//                res.updateConfiguration(conf, dm);
-//                this.recreate();
-//                setResult(Activity.RESULT_OK);
             finish();
         }
     }
@@ -283,41 +111,4 @@ public class SettingsActivity extends PreferenceActivity
                 .registerOnSharedPreferenceChangeListener(this);
     }
 
-    /**
-     * This fragment shows general preferences only. It is used when the
-     * activity is showing a two-pane settings UI.
-     */
-    @TargetApi(Build.VERSION_CODES.HONEYCOMB)
-    public static class GeneralPreferenceFragment extends PreferenceFragment {
-        @Override
-        public void onCreate(Bundle savedInstanceState) {
-            super.onCreate(savedInstanceState);
-            addPreferencesFromResource(R.xml.pref_general);
-        }
-    }
-
-    /**
-     * This fragment shows notification preferences only. It is used when the
-     * activity is showing a two-pane settings UI.
-     */
-    @TargetApi(Build.VERSION_CODES.HONEYCOMB)
-    public static class NotificationPreferenceFragment extends PreferenceFragment {
-        @Override
-        public void onCreate(Bundle savedInstanceState) {
-            super.onCreate(savedInstanceState);
-
-        }
-    }
-
-    /**
-     * This fragment shows data and sync preferences only. It is used when the
-     * activity is showing a two-pane settings UI.
-     */
-    @TargetApi(Build.VERSION_CODES.HONEYCOMB)
-    public static class DataSyncPreferenceFragment extends PreferenceFragment {
-        @Override
-        public void onCreate(Bundle savedInstanceState) {
-            super.onCreate(savedInstanceState);
-        }
-    }
 }