+++ app/src/main/java/org/akvo/caddisfly/util/VerticalScrollView.java
@@ -1,0 +1,76 @@
+/*
+ * Copyright (C) Stichting Akvo (Akvo Foundation)
+ *
+ * This file is part of Akvo Caddisfly
+ *
+ * Akvo Caddisfly is free software: you can redistribute it and modify it under the terms of
+ * the GNU Affero General Public License (AGPL) as published by the Free Software Foundation,
+ * either version 3 of the License or any later version.
+ *
+ * Akvo Caddisfly is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY;
+ * without even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
+ * See the GNU Affero General Public License included below for more details.
+ *
+ * The full license text can also be seen at <http://www.gnu.org/licenses/agpl.html>.
+ */
+
+package org.akvo.caddisfly.util;
+
+import android.content.Context;
+import android.util.AttributeSet;
+import android.util.Log;
+import android.view.MotionEvent;
+import android.widget.ScrollView;
+
+//http://stackoverflow.com/questions/3495890/how-can-i-put-a-listview-into-a-scrollview-without-it-collapsing
+
+public class VerticalScrollView extends ScrollView {
+
+    public VerticalScrollView(Context context) {
+        super(context);
+    }
+
+    public VerticalScrollView(Context context, AttributeSet attrs) {
+        super(context, attrs);
+    }
+
+    public VerticalScrollView(Context context, AttributeSet attrs, int defStyle) {
+        super(context, attrs, defStyle);
+    }
+
+    @Override
+    public boolean onInterceptTouchEvent(MotionEvent ev) {
+        final int action = ev.getAction();
+        switch (action) {
+            case MotionEvent.ACTION_DOWN:
+                Log.i("VerticalScrollview", "onInterceptTouchEvent: DOWN super false");
+                super.onTouchEvent(ev);
+                break;
+
+            case MotionEvent.ACTION_MOVE:
+                return false; // redirect MotionEvents to ourself
+
+            case MotionEvent.ACTION_CANCEL:
+                Log.i("VerticalScrollview", "onInterceptTouchEvent: CANCEL super false");
+                super.onTouchEvent(ev);
+                break;
+
+            case MotionEvent.ACTION_UP:
+                Log.i("VerticalScrollview", "onInterceptTouchEvent: UP super false");
+                return false;
+
+            default:
+                Log.i("VerticalScrollview", "onInterceptTouchEvent: " + action);
+                break;
+        }
+
+        return false;
+    }
+
+    @Override
+    public boolean onTouchEvent(MotionEvent ev) {
+        super.onTouchEvent(ev);
+        Log.i("VerticalScrollview", "onTouchEvent. action: " + ev.getAction());
+        return true;
+    }
+}